Prepare for tap: Change former el-main to single component.
